🚚 api 호출 함수 관리방법 수정
